maze karo/requirements.txt
Flask>=2.0.0
python-dotenv
flask-session==0.4.0